@@ -1,6 +1,6 @@
 {
   "name": "citra-discord-bot",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Citra bot for Discord",
   "author": "chris062689 <chris062689@gmail.com>",
   "preferGlobal": true,
@@ -10,21 +10,32 @@
   "license": "GPL-2.0+",
   "dependencies": {
     "checkenv": "^1.2.2",
-    "discord.js": "^11.3.0",
+    "discord.js": "^12.2.0",
     "ip": "^1.1.5",
-    "logdna": "^1.2.3",
-    "node-schedule": "^1.2.3",
-    "request": "^2.79.0",
-    "request-promise-native": "^1.0.5",
-    "string-similarity": "^1.2.0",
-    "winston": "^2.3.0"
+    "logdna": "^3.5.0",
+    "logdna-winston": "^2.3.1",
+    "node-fetch": "^2.6.0",
+    "node-schedule": "^1.3.2",
+    "string-similarity": "^4.0.1",
+    "typescript": "^3.8.3",
+    "winston": "^3.2.1"
   },
   "devDependencies": {
-    "eslint": "^4.8.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.2.0",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1"
+    "@types/ip": "^1.1.0",
+    "@types/node": "^13.13.4",
+    "@types/node-fetch": "^2.5.7",
+    "@types/string-similarity": "^3.0.0",
+    "@types/ws": "^7.2.4",
+    "eslint": "^6.8.0",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
+    "ts-node": "^8.9.1"
+  },
+  "scripts": {
+    "build": "yarn run tsc",
+    "serve": "yarn run ts-node ./src/server.ts"
   }
 }
