@@ -1,9 +1,10 @@
-const state = require('../state.js');
-const data = require('../data.js');
-const logger = require('../logging.js');
+import state from '../state';
+import * as data from '../data';
+import logger from '../logging';
+import discord = require('discord.js');
 
-exports.roles = ['Admins', 'Moderators'];
-exports.command = function (message) {
+export const roles = ['Admins', 'Moderators'];
+export function command (message: discord.Message) {
   message.mentions.users.map((user) => {
     const count = state.warnings.filter(x => x.id === user.id && !x.cleared);
     if (count != null && count.length > 0) {
@@ -14,7 +15,7 @@ exports.command = function (message) {
       message.channel.send(`${user}, you have no warnings to clear.`);
     }
 
-    logger.info(`${message.author.toString()} has cleared all warnings for ${user.toString()} [${count.length}].`);
+    logger.info(`${message.author.username} has cleared all warnings for ${user} ${user.username} [${count.length}].`);
     state.logChannel.send(`${message.author.toString()} has cleared all warnings for ${user.toString()} [${count.length}].`);
   });
 };
