@@ -1,5 +1,7 @@
-exports.roles = ['Admins', 'Moderators'];
-exports.command = function (message, reply) {
+import discord = require('discord.js');
+
+export const roles = ['Admins', 'Moderators'];
+export function command (message: discord.Message, reply: string) {
   let replyMessage = 'Hello.';
   if (reply == null) {
     replyMessage = message.content.substr(message.content.indexOf(' ') + 1);
@@ -8,4 +10,4 @@ exports.command = function (message, reply) {
   }
 
   message.channel.send(replyMessage);
-};
+}
